requirements.txt
APScheduler==3.11.0
asttokens==3.0.0
atomicwrites==1.4.1
attrs==25.3.0
backtesting==0.6.4
beautifulsoup4==4.13.4
bokeh==3.7.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
comm==0.2.2
config==0.5.1
contourpy==1.3.2
curl_cffi==0.11.1
debugpy==1.8.14
decorator==5.2.1
executing==2.2.0
forex-types==0.0.6
frozendict==2.4.6
idna==3.10
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.7.2
magic-kind==0.2.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
more-itertools==10.7.0
multitasking==0.0.11
narwhals==1.40.0
nest-asyncio==1.6.0
numpy==1.26.3
oanda-candles==0.1.0
oandapyV20==0.6.3
packaging==25.0
pandas==2.2.3
pandas_ta==0.3.14b0
parso==0.8.4
peewee==3.18.1
pillow==11.2.1
platformdirs==4.3.8
plotly==6.1.0
pluggy==0.13.1
prompt_toolkit==3.0.51
protobuf==6.31.0
psutil==7.0.0
pure_eval==0.2.3
py==1.11.0
pycparser==2.22
Pygments==2.19.1
pytest==5.4.3
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
requests==2.32.3
setuptools==80.7.1
six==1.17.0
soupsieve==2.7
stack-data==0.6.3
time-int==0.0.9
tornado==6.5
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
wcwidth==0.2.13
websockets==15.0.1
xyzservices==2025.4.0
yfinance==0.2.61
